crear, consultar y actualizar de registro
