+++ src/main/java/Controlador/Registro/ControladorRegistroUsuario.java
@@ -12,8 +12,7 @@ import Entidades.Preguntas;
 public class ControladorRegistroUsuario {
     EntityManager em;
     Estudiante est;
-    Preguntas pr;
-
+    
     //Estudiante
     public void crearEstudiante (Estudiante estudiante) throws Exception{
         EstudianteJpaController controladorEstudiante = new EstudianteJpaController(Persistence.createEntityManagerFactory("com.mycompany_SianCliente_jar_1.0-SNAPSHOTPU"));
@@ -37,34 +36,4 @@ public class ControladorRegistroUsuario {
         }
         return est;
     }
-    
-    //Preguntas
-    public void crearPreguntas (Preguntas pregunta) throws Exception{
-        PreguntasJpaController controladorPreguntas = new PreguntasJpaController(Persistence.createEntityManagerFactory("com.mycompany_SianCliente_jar_1.0-SNAPSHOTPU"));
-        em = controladorPreguntas.getEntityManager();
-        controladorPreguntas.create(pregunta);
-    }
-    
-    public Preguntas consultarCedulaPreguntas(Estudiante idEstudiante){
-        PreguntasJpaController controladorPreguntas = new PreguntasJpaController(Persistence.createEntityManagerFactory("com.mycompany_SianCliente_jar_1.0-SNAPSHOTPU"));
-        em = controladorPreguntas.getEntityManager();
-        List <Preguntas> lstPreguntas = em.createNamedQuery("Preguntas.findByIdEstudiante",Preguntas.class).setParameter("idestudiante",idEstudiante).getResultList();
-        pr = new Preguntas();
-        if (!lstPreguntas.isEmpty()){
-            pr=lstPreguntas.get(0);
-        }
-        return pr;
-    }
-    
-    public void editarPregunta (Preguntas pregunta) throws Exception{
-        PreguntasJpaController controladorPreguntas = new PreguntasJpaController(Persistence.createEntityManagerFactory("com.mycompany_SianCliente_jar_1.0-SNAPSHOTPU"));
-        em = controladorPreguntas.getEntityManager();
-        controladorPreguntas.edit(pregunta);
-    }
-    
-    public void eliminarPregunta (Estudiante estudiante){
-        
-    }
-    
-    
 }
