+++ src/main/java/JpaControllers/EstudianteJpaController.java
@@ -91,67 +91,67 @@ public class EstudianteJpaController implements Serializable {
             em = getEntityManager();
             em.getTransaction().begin();
             Estudiante persistentEstudiante = em.find(Estudiante.class, estudiante.getIdestudiante());
-            List<Inscripcion> inscripcionListOld = persistentEstudiante.getInscripcionList();
-            List<Inscripcion> inscripcionListNew = estudiante.getInscripcionList();
-            List<Preguntas> preguntasListOld = persistentEstudiante.getPreguntasList();
-            List<Preguntas> preguntasListNew = estudiante.getPreguntasList();
-            List<String> illegalOrphanMessages = null;
-            for (Inscripcion inscripcionListOldInscripcion : inscripcionListOld) {
-                if (!inscripcionListNew.contains(inscripcionListOldInscripcion)) {
-                    if (illegalOrphanMessages == null) {
-                        illegalOrphanMessages = new ArrayList<String>();
-                    }
-                    illegalOrphanMessages.add("You must retain Inscripcion " + inscripcionListOldInscripcion + " since its idestudiante field is not nullable.");
-                }
-            }
-            for (Preguntas preguntasListOldPreguntas : preguntasListOld) {
-                if (!preguntasListNew.contains(preguntasListOldPreguntas)) {
-                    if (illegalOrphanMessages == null) {
-                        illegalOrphanMessages = new ArrayList<String>();
-                    }
-                    illegalOrphanMessages.add("You must retain Preguntas " + preguntasListOldPreguntas + " since its idestudiante field is not nullable.");
-                }
-            }
-            if (illegalOrphanMessages != null) {
-                throw new IllegalOrphanException(illegalOrphanMessages);
-            }
-            List<Inscripcion> attachedInscripcionListNew = new ArrayList<Inscripcion>();
-            for (Inscripcion inscripcionListNewInscripcionToAttach : inscripcionListNew) {
-                inscripcionListNewInscripcionToAttach = em.getReference(inscripcionListNewInscripcionToAttach.getClass(), inscripcionListNewInscripcionToAttach.getIdinscripcion());
-                attachedInscripcionListNew.add(inscripcionListNewInscripcionToAttach);
-            }
-            inscripcionListNew = attachedInscripcionListNew;
-            estudiante.setInscripcionList(inscripcionListNew);
-            List<Preguntas> attachedPreguntasListNew = new ArrayList<Preguntas>();
-            for (Preguntas preguntasListNewPreguntasToAttach : preguntasListNew) {
-                preguntasListNewPreguntasToAttach = em.getReference(preguntasListNewPreguntasToAttach.getClass(), preguntasListNewPreguntasToAttach.getIdpreguntas());
-                attachedPreguntasListNew.add(preguntasListNewPreguntasToAttach);
-            }
-            preguntasListNew = attachedPreguntasListNew;
-            estudiante.setPreguntasList(preguntasListNew);
+//            List<Inscripcion> inscripcionListOld = persistentEstudiante.getInscripcionList();
+//            List<Inscripcion> inscripcionListNew = estudiante.getInscripcionList();
+//            List<Preguntas> preguntasListOld = persistentEstudiante.getPreguntasList();
+//            List<Preguntas> preguntasListNew = estudiante.getPreguntasList();
+//            List<String> illegalOrphanMessages = null;
+//            for (Inscripcion inscripcionListOldInscripcion : inscripcionListOld) {
+//                if (!inscripcionListNew.contains(inscripcionListOldInscripcion)) {
+//                    if (illegalOrphanMessages == null) {
+//                        illegalOrphanMessages = new ArrayList<String>();
+//                    }
+//                    illegalOrphanMessages.add("You must retain Inscripcion " + inscripcionListOldInscripcion + " since its idestudiante field is not nullable.");
+//                }
+//            }
+//            for (Preguntas preguntasListOldPreguntas : preguntasListOld) {
+//                if (!preguntasListNew.contains(preguntasListOldPreguntas)) {
+//                    if (illegalOrphanMessages == null) {
+//                        illegalOrphanMessages = new ArrayList<String>();
+//                    }
+//                    illegalOrphanMessages.add("You must retain Preguntas " + preguntasListOldPreguntas + " since its idestudiante field is not nullable.");
+//                }
+//            }
+//            if (illegalOrphanMessages != null) {
+//                throw new IllegalOrphanException(illegalOrphanMessages);
+//            }
+//            List<Inscripcion> attachedInscripcionListNew = new ArrayList<Inscripcion>();
+//            for (Inscripcion inscripcionListNewInscripcionToAttach : inscripcionListNew) {
+//                inscripcionListNewInscripcionToAttach = em.getReference(inscripcionListNewInscripcionToAttach.getClass(), inscripcionListNewInscripcionToAttach.getIdinscripcion());
+//                attachedInscripcionListNew.add(inscripcionListNewInscripcionToAttach);
+//            }
+//            inscripcionListNew = attachedInscripcionListNew;
+//            estudiante.setInscripcionList(inscripcionListNew);
+//            List<Preguntas> attachedPreguntasListNew = new ArrayList<Preguntas>();
+//            for (Preguntas preguntasListNewPreguntasToAttach : preguntasListNew) {
+//                preguntasListNewPreguntasToAttach = em.getReference(preguntasListNewPreguntasToAttach.getClass(), preguntasListNewPreguntasToAttach.getIdpreguntas());
+//                attachedPreguntasListNew.add(preguntasListNewPreguntasToAttach);
+//            }
+//            preguntasListNew = attachedPreguntasListNew;
+//            estudiante.setPreguntasList(preguntasListNew);
             estudiante = em.merge(estudiante);
-            for (Inscripcion inscripcionListNewInscripcion : inscripcionListNew) {
-                if (!inscripcionListOld.contains(inscripcionListNewInscripcion)) {
-                    Estudiante oldIdestudianteOfInscripcionListNewInscripcion = inscripcionListNewInscripcion.getIdestudiante();
-                    inscripcionListNewInscripcion.setIdestudiante(estudiante);
-                    inscripcionListNewInscripcion = em.merge(inscripcionListNewInscripcion);
-                    if (oldIdestudianteOfInscripcionListNewInscripcion != null && !oldIdestudianteOfInscripcionListNewInscripcion.equals(estudiante)) {
-                        oldIdestudianteOfInscripcionListNewInscripcion.getInscripcionList().remove(inscripcionListNewInscripcion);
-                        oldIdestudianteOfInscripcionListNewInscripcion = em.merge(oldIdestudianteOfInscripcionListNewInscripcion);
-                    }
-                }
-            }
-            for (Preguntas preguntasListNewPreguntas : preguntasListNew) {
-                if (!preguntasListOld.contains(preguntasListNewPreguntas)) {
-                    Estudiante oldIdestudianteOfPreguntasListNewPreguntas = preguntasListNewPreguntas.getIdestudiante();
-                    preguntasListNewPreguntas.setIdestudiante(estudiante);
-                    preguntasListNewPreguntas = em.merge(preguntasListNewPreguntas);
-                    if (oldIdestudianteOfPreguntasListNewPreguntas != null && !oldIdestudianteOfPreguntasListNewPreguntas.equals(estudiante)) {
-                        oldIdestudianteOfPreguntasListNewPreguntas.getPreguntasList().remove(preguntasListNewPreguntas);
-                        oldIdestudianteOfPreguntasListNewPreguntas = em.merge(oldIdestudianteOfPreguntasListNewPreguntas);
-                    }
-                }
-            }
+//            for (Inscripcion inscripcionListNewInscripcion : inscripcionListNew) {
+//                if (!inscripcionListOld.contains(inscripcionListNewInscripcion)) {
+//                    Estudiante oldIdestudianteOfInscripcionListNewInscripcion = inscripcionListNewInscripcion.getIdestudiante();
+//                    inscripcionListNewInscripcion.setIdestudiante(estudiante);
+//                    inscripcionListNewInscripcion = em.merge(inscripcionListNewInscripcion);
+//                    if (oldIdestudianteOfInscripcionListNewInscripcion != null && !oldIdestudianteOfInscripcionListNewInscripcion.equals(estudiante)) {
+//                        oldIdestudianteOfInscripcionListNewInscripcion.getInscripcionList().remove(inscripcionListNewInscripcion);
+//                        oldIdestudianteOfInscripcionListNewInscripcion = em.merge(oldIdestudianteOfInscripcionListNewInscripcion);
+//                    }
+//                }
+//            }
+//            for (Preguntas preguntasListNewPreguntas : preguntasListNew) {
+//                if (!preguntasListOld.contains(preguntasListNewPreguntas)) {
+//                    Estudiante oldIdestudianteOfPreguntasListNewPreguntas = preguntasListNewPreguntas.getIdestudiante();
+//                    preguntasListNewPreguntas.setIdestudiante(estudiante);
+//                    preguntasListNewPreguntas = em.merge(preguntasListNewPreguntas);
+//                    if (oldIdestudianteOfPreguntasListNewPreguntas != null && !oldIdestudianteOfPreguntasListNewPreguntas.equals(estudiante)) {
+//                        oldIdestudianteOfPreguntasListNewPreguntas.getPreguntasList().remove(preguntasListNewPreguntas);
+//                        oldIdestudianteOfPreguntasListNewPreguntas = em.merge(oldIdestudianteOfPreguntasListNewPreguntas);
+//                    }
+//                }
+//            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
